GN-4323: Move functions to util files
